poem_app_backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.0.2
bcrypt==4.0.1
PyJWT==2.8.0
Pillow==10.0.1
requests==2.31.0
gunicorn==21.2.0 